Update to wgpu 0.6
We've been stuck on wgpu 0.4 for quite a while due to the new lifetimes
introduced around the render pass. By collecting all the draw calls into
a Vec first and then later building the render pass, we can safely
borrow all the data we need. In a way this also makes sense, because the
way livesplit-core's renderer currently works, allows it to drop meshes
before we ever submit any of the draw calls into the queue. The Vec
keeps them alive long enough so we can submit them.

Additionally this introduces the usage of the wgpu-mipmap crate that
generates the mip maps for all the textures. Previously our wgpu
renderer did not generate any mip maps at all.