Remove wgpu and switch to software renderer
Turns out that it's very fast to render with tiny-skia. In fact it's
much faster than using the GPU and consumes a lot less RAM. Additionally
it removes all sorts of problems you would encounter with all the
different graphics APIs, their differences and driver bugs. For now this
is the most optimal renderer.