Add Support for Auto Splitters
This adds support for specifying and loading an auto splitter. There's
also some other minor improvements such as using lld for linking on
Windows, using mimalloc as the global allocator and fixing a bug where a
minimized window would crash the application. Additionally this also
increases the chunk size when transposing the colors, allowing for
better code generation when AVX is active.