Optimize Debug Compile Times
We previously bumped everything to `opt-level = 3`, which is completely
overkill. We really only need `tiny-skia`, which is our renderer, and
the `transmute` function, which processes the whole window on every
frame, to be fast. While this definitely slows down the performance a
little bit overall, the wins in compile times are huge and if you want a
fully optimized build, you should be using the release build anyway.